- upg to golang-lru/v2 - track close errors in metrics - move BlockstoreCloser and readCloser to utils - add comment for sleep in test
